+++ HISTORY.md
@@ -5,6 +5,7 @@
     * log pen vectors
     * export pen trails as SVG
     * access pen trails as SVG_Costume: new "pen vectors" reporter variant of "pen trails"
+    * new Snap! API: programmatically broadcast messages and optionally wait from outside Snap!
 * **Notable Changes:**
     * when creating a costume from pen trails (raster or vector) make its rotation center the position of the sprite
 * **Notable Fixes:**
@@ -12,6 +13,9 @@
 * **Translation Updates:**
     * NEW Slovak translation, thanks, Peter Lukacovic
     * German
+
+### 2019-12-15
+* gui, threads:  new Snap! API: programmatically broadcast messages and optionally wait from outside Snap!
 
 ### 2019-12-13
 * added direct relabelling option to pen trails blocks' context menus